Update pmod_perturb.py
A utility script relied upon in this block of code in LOADGF/LN is named 'prepare_planet_model.py', however, this block of code calls a function 'prepare_earth_model.py' which does not exist. I have altered the utility script name written here to be consistent with the existing function in LOADGF/LN. 

Similarly, I have changed variable names in this script from 'earth_model', 'earth_mass', etc. to 'planet_model', 'planet_mass', etc. to be consistent with preexisting naming conventions in LoadDef. 

I have also added a note for users to read that indicates that some of the default values in the function prepare_planet_model.py are specific to Earth and should be altered if they are working on a different planetary body.